Add files via upload
Hello friends! I have been assigned a task to develop a simple plugin to work with a WooCommerce shopping cart. I have never coded a plugin before, and am struggling to get it to activate. So far, I have developed the site, created the store, created a plugin but the code is off. Can anyone tell me what I am doing wrong? I am looking everywhere for a snippet of the correct code but I can't find it anywhere.

My task is as follows: Create a plugin for WooCommerce (https://docs.woocommerce.com/document/create-a-plugin/), that will display an alert box when you change the quantity of an item in the cart with the following text:

"You just changed the quantity of ABC to 2", where ABC is the SKU name, and 2 is the current (new) quantity of the item.

this is the error i am getting when I click activate: Parse error: syntax error, unexpected end of file in /home/jdndkvccq40f/public_html/wp-content/plugins/capetown/my-plugin-project.php on line 58

This is what I have done so far:

the name of the plugin is capetown.php